Renaming queues module to pika
